test: succes test for GetMenu route
Implemented tests for menu handler GetMenu using mtest and mock
database.

+++ internal/menu/handler.go
@@ -132,7 +132,7 @@ func CreateMenuItem(client db.IMongoClient) gin.HandlerFunc {
 		item := MenuItem{
 			Name:        name,
 			Description: description,
-			Price:       float32(priceFloat),
+			Price:       float64(priceFloat),
 			Category:    category,
 			Img:         imagePath,
 		}
