+++ internal/order/model.go
@@ -11,7 +11,7 @@ import (
 type Order struct {
 	ID          primitive.ObjectID `bson:"_id,omitempty"       json:"id"`
 	Items       []menu.MenuItem    `bson:"items"               json:"items"       validate:"required"`
-	TotalPrice  float32            `bson:"totalPrice"          json:"totalPrice"`
+	TotalPrice  float64            `bson:"totalPrice"          json:"totalPrice"`
 	TableNumber int                `bson:"tableNumber"         json:"tableNumber"`
 	IsClosed    bool               `bson:"isClosed"            json:"isClosed"`
 	ServedAt    *time.Time         `bson:"servedAt,omitempty"  json:"servedAt"`
